--- a/package.json
+++ b/package.json
@@ -3,6 +3,8 @@
   "license": "MIT",
   "dependencies": {
     "@stdlib/math-base-special-betainc": "^0.0.6",
+    "@tensorflow/tfjs-backend-cpu": "^3.14.0",
+    "@tensorflow/tfjs-core": "^3.14.0",
     "three": "^0.128.0"
   },
   "devDependencies": {