--- a/scripts/generate_sample_data.r
+++ b/scripts/generate_sample_data.r
@@ -22,3 +22,22 @@ writeLines(
   jsonlite::toJSON(unclass(p$x), auto_unbox = TRUE, null = "null"),
   "../dev/display_scatter_3d/static/sample_data.json"
 )
+
+set.seed(1)
+
+p <- detour(tourr::flea, tour_aes(projection = where(is.numeric), color = species, label = species)) |>
+  display_sage()
+
+writeLines(
+  jsonlite::toJSON(unclass(p$x), auto_unbox = TRUE, null = "null"),
+  "../dev/display_sage_2d/static/sample_data.json"
+)
+
+p <- detour(tourr::flea, tour_aes(projection = where(is.numeric), color = species, label = species)) |>
+  tour_path(grand_tour(3)) |>
+  display_sage()
+
+writeLines(
+  jsonlite::toJSON(unclass(p$x), auto_unbox = TRUE, null = "null"),
+  "../dev/display_sage_3d/static/sample_data.json"
+)