--- a/tsconfig.json
+++ b/tsconfig.json
@@ -1,20 +1,20 @@
 {
-    "compilerOptions": {
-        "noImplicitAny": true,
-        "suppressImplicitAnyIndexErrors": true,
-        "module": "commonjs",
-        "sourceMap": true,
-        "esModuleInterop": true,
-        "resolveJsonModule": true,
-        "target": "es5",
-        "allowJs": false,
-        "lib": [
-            "dom",
-            "es6",
-            "es2019"
-        ]
-    },
-    "include": [
-        "srcts/"
+  "compilerOptions": {
+    "noImplicitAny": true,
+    "suppressImplicitAnyIndexErrors": true,
+    "module": "commonjs",
+    "sourceMap": true,
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "target": "es5",
+    "allowJs": false,
+    "lib": [
+      "dom",
+      "es6",
+      "es2019"
     ]
+  },
+  "include": [
+    "srcts/"
+  ]
 }