--- a/webpack/dev.config.ts
+++ b/webpack/dev.config.ts
@@ -8,10 +8,8 @@ module.exports = {
       "./srcts/display_scatter_2d/index.ts",
     "display_scatter_3d/display_scatter_3d":
       "./srcts/display_scatter_3d/index.ts",
-    "display_sage_2d/display_sage_2d":
-      "./srcts/display_sage_2d/index.ts",
-    "display_sage_3d/display_sage_3d":
-      "./srcts/display_sage_3d/index.ts",
+    "display_sage_2d/display_sage_2d": "./srcts/display_sage_2d/index.ts",
+    "display_sage_3d/display_sage_3d": "./srcts/display_sage_3d/index.ts",
   },
   module: {
     rules: [
@@ -40,22 +38,32 @@ module.exports = {
   },
   plugins: [
     new CopyPlugin({
-      patterns: [{
-        from: "dev/display_scatter_2d/static/",
-        to: "display_scatter_2d",
-      }, {
-        from: "dev/display_scatter_3d/static/",
-        to: "display_scatter_3d",
-      }, {
-        from: "dev/display_sage_2d/static/",
-        to: "display_sage_2d",
-      }, {
-        from: "dev/display_sage_3d/static/",
-        to: "display_sage_3d",
-      }, {
-        from: "dev/index.html",
-        to: "",
-      }],
+      patterns: [
+        {
+          from: "dev/display_scatter_2d/static/",
+          to: "display_scatter_2d",
+        },
+        {
+          from: "dev/display_scatter_2d/static/",
+          to: "display_scatter_2d",
+        },
+        {
+          from: "dev/display_scatter_3d/static/",
+          to: "display_scatter_3d",
+        },
+        {
+          from: "dev/display_sage_2d/static/",
+          to: "display_sage_2d",
+        },
+        {
+          from: "dev/display_sage_3d/static/",
+          to: "display_sage_3d",
+        },
+        {
+          from: "dev/index.html",
+          to: "",
+        },
+      ],
     }),
   ],
   devServer: {