--- a/webpack/package.config.ts
+++ b/webpack/package.config.ts
@@ -4,10 +4,10 @@ const webpack = require("webpack");
 module.exports = {
   mode: "production",
   entry: {
-    "display_scatter_2d": "./srcts/display_scatter_2d/index.ts",
-    "display_scatter_3d": "./srcts/display_scatter_3d/index.ts",
-    "display_sage_2d": "./srcts/display_sage_2d/index.ts",
-    "display_sage_3d": "./srcts/display_sage_3d/index.ts",
+    display_scatter_2d: "./srcts/display_scatter_2d/index.ts",
+    display_scatter_3d: "./srcts/display_scatter_3d/index.ts",
+    display_sage_2d: "./srcts/display_sage_2d/index.ts",
+    display_sage_3d: "./srcts/display_sage_3d/index.ts",
   },
   module: {
     rules: [
@@ -33,8 +33,9 @@ module.exports = {
         vendor: {
           test: /[\\/]node_modules[\\/]/,
           name(module) {
-            const packageName =
-              module.context.match(/[\\/]node_modules[\\/](.*?)([\\/]|$)/)[1];
+            const packageName = module.context.match(
+              /[\\/]node_modules[\\/](.*?)([\\/]|$)/
+            )[1];
             return packageName.replace("@", "");
           },
           chunks: "all",