… dev/display_sage_2d/static/sample_data.json
… dev/display_sage_3d/static/sample_data.json
… dev/display_scatter_2d/static/sample_data.json
… dev/display_scatter_3d/static/sample_data.json
… inst/htmlwidgets/lib/display_sage_2d.bundle.js
… inst/htmlwidgets/lib/display_sage_3d.bundle.js
… inst/htmlwidgets/lib/display_scatter_2d.bundle.js
… inst/htmlwidgets/lib/display_scatter_3d.bundle.js
… inst/htmlwidgets/lib/long.bundle.js
… inst/htmlwidgets/lib/runtime.bundle.js
… inst/htmlwidgets/lib/seedrandom.bundle.js
… inst/htmlwidgets/lib/stdlib.bundle.js
… inst/htmlwidgets/lib/tensorflow.bundle.js
… inst/htmlwidgets/lib/tensorflow.bundle.js.LICENSE.txt
… yarn.lock