closes #94

- use tensorflow `Tensor2D` as the data type for dataset and projection matrices
- remove hand-coded projection logic and replace with tensorflow `matMul`
- Modify functions like `project()` to return flat `Float32Arrays` instead of nested arrays for improved performance

A performance comparison for the 3D `display_scatter()` tour. Before:
<img width="453" alt="before-tfjs" src="https://user-images.githubusercontent.com/39182232/158927022-ce9878a2-8a6d-45c5-8437-ca70050e5e5a.png">

And after implementing the logic described above:
<img width="453" alt="after-tfjs" src="https://user-images.githubusercontent.com/39182232/158927060-6107662c-eb3c-4f97-9c1d-6988fe99cfd9.png">

From the above, the scripting time is reduced from 1783ms to 704 ms when profiling across a ~10s window. This is an improvement of around 60%, or around 50% if you include rendering, painting, etc.

This implementation with tensorflow currently only uses the `CPU` backend, and performance can be improved further by using the `wasm` backend. However I haven't managed to get this to work with `HTMLWidgets` yet, as there are issues importing the `.wasm` files.

I would expect using the `wasm` backend will bring the scripting time down from ~700ms to ~500ms, which I've observed running it locally.